requirements.txt
discord.py
requests==2.3.0